fuzzymunge: simplify files on command line
The whole list of arbitrary input and output files was nice
and flexible, but it made simple use somewhat of a pain.
Instead, let's provide sed-style options: either you dump
input files to stdout, or you modify files in-place. That
gives us a nice clean:

  fuzzydups -b base <in.list >out.list

or

  fuzzydups -i *.list <base

for the most common use cases.