build: add a target to install man-pages
MFM's upstream install ruleset learned about this recently,
so we can make us of it trivially. We also add an
"install-doc" rule that chains to "install-man".

We could drop our "doc" and "install-doc" aliases and just
use "man" as the MFM ruleset does. But this keeps the door
open to building and installing HTML documentation, too, and
the user need only care about the "doc" target.